requirements.txt
pycoingecko==1.3.0
click
etherscan-python==2.0.2
black==21.6b0
isort==5.8.0
mypy==0.910
pylint==2.8.2
pytest-mock==3.6.1
pyyaml==5.4.1
